+++ libs/obs/include/mrpt/obs/CObservationRotatingScan.h
@@ -21,7 +21,6 @@ namespace mrpt::obs
 {
 class CObservationVelodyneScan;
 class CObservation2DRangeScan;
-class CObservationPointCloud;
 
 /** \addtogroup mrpt_obs_grp
  * @{ */
@@ -33,11 +32,6 @@ class CObservationPointCloud;
  * detection directly on "range images" and on points stored as a matrix in the
  * member organizedPoints.
  *
- * This class can also import data from KITTI dataset-like binary files
- * containing unorganized (non "undistorted", i.e. without compensation for
- * lidar motion) point clouds, which get organized into a 2D range image for
- * easier filtering and postprocessing.
- *
  * Check out the main data fields in the list of members below.
  *
  *  Note that this object has \b two timestamp fields:
@@ -48,14 +42,10 @@ class CObservationPointCloud;
  * local computer-based timestamp based on the reception of the message in
  * the computer.
  *
- *  Both timestamps correspond to the firing of the <b>first</b> laser in
- * the <b>first</b> CObservationRotatingScan::scan_packets packet.
+ * Both timestamps correspond to the firing of the **first** laser in
+ * the scan, i.e. the first column in organizedPoints.
  *
- *  <div align=center> <img src="velodyne_axes.jpg"> </div>
- *
- * API for accurate reconstruction of point clouds from raw range images:
- *  - generatePointCloud()
- *  - generatePointCloudAlongSE3Trajectory()
+ * The reference frame for the 3D LIDAR is with +X pointing forward, +Z up.
  *
  * \note New in MRPT 2.0.0
  * \sa CObservation, mrpt::hwdrivers::CVelodyneScanner
@@ -65,12 +55,25 @@ class CObservationRotatingScan : public CObservation
 	DEFINE_SERIALIZABLE(CObservationRotatingScan, mrpt::obs)
 
    public:
+	enum class ExternalStorageFormat : uint8_t
+	{
+		None = 0,  //!< is always stored in memory
+		MRPT_Serialization,	 //!< Uses mrpt-serialization binary file
+		/// Plain text, format explained in saveToTextFile()
+		PlainTextFile
+	};
+
+   protected:
+	ExternalStorageFormat m_externally_stored = ExternalStorageFormat::None;
+	std::string m_external_file;
+
+   public:
 	/** @name Scan range data
 		@{ */
 
 	/** Number of "Lidar rings" (e.g. 16 for a Velodyne VLP16, etc.). This
 	 * should be constant for a given LiDAR scanner.
-	 * All matrices in `imageLayer_*` have this number of rows.
+	 * All matrices defined below have this number of rows.
 	 */
 	uint16_t rowCount{0};
 
@@ -156,18 +159,54 @@ class CObservationRotatingScan : public CObservation
 	void fromScan2D(const mrpt::obs::CObservation2DRangeScan& o);
 
 	/** Will convert from another observation if it's any of the supported
-	 * source types (see fromVelodyne(), fromScan2D(), fromPointCloud()) and
+	 * source types (see fromVelodyne(), fromScan2D()) and
 	 * return true, or will return false otherwise if there is no known way to
 	 * convert from the passed object. */
 	bool fromGeneric(const mrpt::obs::CObservation& o);
-
 	/** @} */
 
-	/** @name "Convert to" API
+	/** @name Delayed-load manual control methods.
+		@{ */
+	// See base class docs.
+	void load() const override;
+	void unload() const override;
+	/** @} */
+
+	/** \name Point cloud external storage functions
 	 * @{ */
+	inline bool isExternallyStored() const
+	{
+		return m_externally_stored != ExternalStorageFormat::None;
+	}
+	inline const std::string& getExternalStorageFile() const
+	{
+		return m_external_file;
+	}
+	void setAsExternalStorage(
+		const std::string& fileName, const ExternalStorageFormat fmt);
+
+	void overrideExternalStorageFormatFlag(const ExternalStorageFormat fmt)
+	{
+		m_externally_stored = fmt;
+	}
+
+	/** Write scan data to a plain text, each line has:
+	 *   `x y z intensity range row_idx col_idx`
+	 *
+	 * For each point in the organized point cloud.
+	 * Invalid points (e.g. no lidar return) are stored as (x,y,z)=(0,0,0) and
+	 * range=0.
+	 *
+	 * \return true on success
+	 */
+	bool saveToTextFile(const std::string& filename) const;
+
+	/** Loads the range, intensity, and organizedPoints members from a plain
+	 * text file in the format describd in saveToTextFile()
+	 */
+	bool loadFromTextFile(const std::string& filename);
 
 	/** @} */
-
 	// See base class docs
 	mrpt::system::TTimeStamp getOriginalReceivedTimeStamp() const override;